@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.4 || ^8.0",
         "ext-json": "*",
         "ext-pdo": "*",
         "ext-zlib": "*",
@@ -21,7 +21,7 @@
     },
     "require-dev": {
         "symplify/easy-coding-standard": "^11",
-        "phpunit/phpunit": "^7"
+        "phpunit/phpunit": "^9"
     },
     "autoload": {
         "psr-4": {
